Update index.html
Uncaught TypeError: adj is undefined
    generatePair https://robertecurtin.github.io/chatGpt.AdjectiveNounPair/:27
    onclick https://robertecurtin.github.io/chatGpt.AdjectiveNounPair/:1

It looks like the issue might be with the getAdjectiveStartingWith function. If there are no adjectives in the list that start with the specified letter, then the function won't return anything, which could result in the adj variable being undefined in the generatePair function.

To fix this issue, we can modify the getAdjectiveStartingWith function to return a default adjective if there are no adjectives in the list that start with the specified letter. Here's an updated version of the code:

...

In this version of the code, we've added a defaultAdj variable that is set to 'awesome'. If there are no adjectives in the list that start with the specified letter, the getAdjectiveStartingWith function will return the defaultAdj variable. This ensures that adj will always be defined in the generatePair function.