Update index.html
This has a bug, inputing a capitalized noun does not find an adjective from the list

You're right, the bug is that the function getAdjectiveStartingWith() is using the lowercase version of the first letter of the noun to filter the adjective list. To fix this, we can simply lowercase the first letter of the noun before passing it to getAdjectiveStartingWith(). Here's the updated code:
...